+++ core/src/nextpas.core.js.pure.base.pas
@@ -181,6 +181,20 @@ begin
   SetLength(Heap, 0);
 end;
 
+function JsCategoryFromErrorCategory(const ACategory: TErrorCategory): TJsErrorCategory; inline;
+begin
+  case ACategory of
+    ecParse: Result := jecSyntax;
+    ecNullReference: Result := jecReference;
+    ecInvalidArgument, ecInvalidOperation: Result := jecType;
+    ecNotImplemented, ecNotSupported: Result := jecNotSupported;
+    ecTimeout: Result := jecTimeout;
+    ecResourceExhausted: Result := jecMemory;
+    ecInternal: Result := jecUnknown;
+    else Result := jecUnknown;
+  end;
+end;
+
 function JsPureCall(ACtx: IJsContext; const Hosts: TJsPureHostArray; const AFunc, AThis: TJsValue; const AArgs: array of TJsValue; ABackend: TJsBackendKind): TJsValue; inline;
 var LIdx: Integer; LName: string;
 begin
@@ -191,9 +205,9 @@ begin
   LIdx := JsPureFindHost(Hosts, LName);
   if LIdx < 0 then Exit;
   case Hosts[LIdx].Kind of
-    0: try Result := Hosts[LIdx].Func(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend); end;
-    1: try Result := Hosts[LIdx].Method(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend); end;
-    2: try Result := Hosts[LIdx].Proc(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend); end;
+    0: try Result := Hosts[LIdx].Func(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend); end;
+    1: try Result := Hosts[LIdx].Method(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend); end;
+    2: try Result := Hosts[LIdx].Proc(ACtx, AThis, AArgs); except on E: EJsError do raise; on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend); on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend); end;
   end;
 end;
 
@@ -299,8 +313,8 @@ begin
               if LHasArg then Result := LHandler(ACtx, LThis, LSingle) else Result := LHandler(ACtx, LThis, LNoArgs);
             except
               on E: EJsError do raise;
-              on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend);
-              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend);
+              on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend);
+              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend);
             end;
             Exit;
           end;
@@ -311,8 +325,8 @@ begin
               if LHasArg then Result := LMethod(ACtx, LThis, LSingle) else Result := LMethod(ACtx, LThis, LNoArgs);
             except
               on E: EJsError do raise;
-              on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend);
-              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend);
+              on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend);
+              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend);
             end;
             Exit;
           end;
@@ -323,8 +337,8 @@ begin
               if LHasArg then Result := LProc(ACtx, LThis, LSingle) else Result := LProc(ACtx, LThis, LNoArgs);
             except
               on E: EJsError do raise;
-              on E: ENextPasError do raise EJsError.Create(E.Message, jecUnknown, 'Error', '', ABackend);
-              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, 'Error', '', ABackend);
+              on E: ENextPasError do raise EJsError.Create(E.Message, JsCategoryFromErrorCategory(E.Category), E.ClassName, '', ABackend);
+              on E: TObject do raise EJsError.Create(E.ClassName, jecUnknown, E.ClassName, '', ABackend);
             end;
             Exit;
           end;